==> 09_pre-post-CURE-exploration.Rmd <==
---
title: "09_pre-post-CURE-exploration"
author: "Deidre Jaeger"
date: "3/5/2020"
output: pdf_document
---

```{r global_options, include=FALSE, echo=FALSE}
knitr::opts_chunk$set(warning=FALSE, message=FALSE, results= FALSE, echo= TRUE)

# echo = false to turn off code

```


```{r load-libraries, echo = FALSE}
library(ggplot2)
library(dplyr)
library(tidyr) #for pivot
library(lubridate) # for dates
library(agricolae) # for the tukey test
```


```{r load-data, echo = FALSE}

# pre <- read.csv(file = "data/student-data/Pre-assessment EBIO 1250- section 012_March 5, 2020_10.19.csv",
#                 header = T,
#                 stringsAsFactors = F)
# 
# post <- read.csv(file = "data/student-data/post-assessment EBIO 1250- section 012_March 5, 2020_10.19.csv",
#                  header = T,
#                  stringsAsFactors = F)

pre <- read.csv(file = "data/student-data/Pre-assessment2 EBIO 1250- section 012_March 5, 2020_17.33.csv",
                header = T,
                stringsAsFactors = F,
                na.strings=c("","NA"))

post <- read.csv(file = "data/student-data/post-assessment2 EBIO 1250- section 012_March 5, 2020_17.33.csv",
                 header = T,
                 stringsAsFactors = F,
                 na.strings=c("","NA"))


```

```{r clean-fuse-datasets, echo = FALSE}
# add a type column
pre$type <- "pre2019"
post$type <- "post2019"

# remove the first 2 rows in each dataset that dont contain variables
pre <- pre[-c(1:2),]
post <- post[-c(1:2),]


colnames(post) == colnames(pre)


# poking around
ncol(post)
which(colnames(post)=="Q24" )
which(colnames(post)=="Q25" )
colnames(pre[61])

# remove consent questions
post2 <- post[, -c(which(colnames(post)=="Q24" ), which(colnames(post)=="Q25" ))]

# remove weird parent topics questions
pre2 <- pre[, -c(which(colnames(pre)== colnames(pre[61])),which(colnames(pre)==colnames(pre[62]) ))]

# check that column names the same
colnames(post2) == colnames(pre2)

# Fuse the datasets
pre.post <- rbind(pre2, post2)


# clean up classes being coded as characters
str(pre.post)

# change columns of quantitative questions to numeric
pre.post[, (which(colnames(pre.post)=="Q8.Q13_1" ):which(colnames(pre.post)=="Q34.40_7"))] <- sapply(pre.post[, (which(colnames(pre.post)=="Q8.Q13_1" ):which(colnames(pre.post)=="Q34.40_7"))], as.numeric)

# Check structure
str(pre.post)

# change factor order
pre.post$type <- factor(pre.post$type, levels = c("pre2019","post2019"))

str(pre.post$type)

```
Apple Biology
It is likely/unlikely that….
Q8. Apple trees can grow from seeds as well as from attaching (grafting) a clipped branch to the base of another tree.
Minor improvement

Q9. There are more than 2,500 kinds of known apple trees presently growing in the United States.
Full improvement
Q10. Most apple trees are self-incompatible, so you need two apple trees of the exact same kind to promote outcrossing.

Improvement
Q11. Fallen fruits can lead to human/wildlife conflicts and often require city governing policies.

Minorly regressive
Q12. Apple trees should be pruned to reduce disease and promote more fruit production.
Minor improvement
Q13. Pollinators always have a positive effect on apple trees.
Minorly regressive


```{r t-tests-apple-biology, results = TRUE}
# Q8. Apple trees can grow from seeds as well as from attaching (grafting) a clipped branch to the base of another tree.
# Q8 _ not sig
t.test(pre.post$Q8.Q13_1 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_1 ~ pre.post$type)

# Q9. There are more than 2,500 kinds of known apple trees presently growing in the United States.
# Q9 -  not  sig
t.test(pre.post$Q8.Q13_2 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_2 ~ pre.post$type)

# Q10. Most apple trees are self-incompatible, so you need two apple trees of the exact same kind to promote outcrossing.
# Q10 _ not sig
t.test(pre.post$Q8.Q13_3 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_3 ~ pre.post$type)

# Q11. Fallen fruits can lead to human/wildlife conflicts and often require city governing policies.
# Q11 _ not sig
t.test(pre.post$Q8.Q13_4 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_4 ~ pre.post$type)

# Q12. Apple trees should be pruned to reduce disease and promote more fruit production.
# Q12 _ not sig, but less variation in the post test
t.test(pre.post$Q8.Q13_5 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_5 ~ pre.post$type)

# Q13. Pollinators always have a positive effect on apple trees.
# Q13 _ not sig
t.test(pre.post$Q8.Q13_6 ~ pre.post$type)
boxplot(pre.post$Q8.Q13_6 ~ pre.post$type)

```

Q14. The photo above shows two branch section of tree leaves with a letter "A" and "B" labeled below each branch. Which set of leaves is most likely from an apple tree? 
Multiple choices: A, B, A & B, Neither A or B

```{r apple-photo-id, results = TRUE}

# Q14 _ not sig
t.test(pre.post$Q14 ~ pre.post$type)
boxplot(pre.post$Q14 ~ pre.post$type)

```


Place-based Ecology
It is likely/unlikely that….
Q15. Most apple trees in Boulder were established through natural dispersal
Q16. Apple tree survival in Boulder is most threatened by humans who want different kinds of tasting apples
Q17. 6,000-19,000 lbs of apple waste may go to waste each year if Community Fruit Rescue didn’t lead community harvests
Q18. Many of the oldest apple trees in Boulder are 100-150 years old and close to the end of their lifespan.
Q19. Fire blight is a disease management concern for fruit trees on the Front Range. 



```{r place-based-ecology, results = TRUE}

# Q15. Most apple trees in Boulder were established through natural dispersal
# Q15 _ not sig
t.test(pre.post$Q15.19_1 ~ pre.post$type)
boxplot(pre.post$Q15.19_1 ~ pre.post$type)

# Q16. Apple tree survival in Boulder is most threatened by humans who want different kinds of tasting apples
# Q16 _ significant diff
t.test(pre.post$Q15.19_2 ~ pre.post$type)
boxplot(pre.post$Q15.19_2 ~ pre.post$type)

# Q17. 6,000-19,000 lbs of apple waste may go to waste each year if Community Fruit Rescue didn’t lead community harvests
# Q17 _ not sig
# t.test(pre.post$Q15.19_3 ~ pre.post$type)
boxplot(pre.post$Q15.19_3 ~ pre.post$type)

# Q18. Many of the oldest apple trees in Boulder are 100-150 years old and close to the end of their lifespan.
# Q18 _ significant difference
t.test(pre.post$Q15.19_4 ~ pre.post$type)
boxplot(pre.post$Q15.19_4 ~ pre.post$type)

# Q19. Fire blight is a disease management concern for fruit trees on the Front Range. 
# Q19 _ signficant difference
# t.test(pre.post$Q15.19_5 ~ pre.post$type)
boxplot(pre.post$Q15.19_5 ~ pre.post$type)

```

Non-content questions
(SALG) Presently I Can…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q20. Find articles relevant to a particular problem in professional journals or elsewhere
Q21. Critically read articles about issues raised in class
Q22. Identify patterns in data
Q23. Recognize a sound argument and appropriate use of evidence
Q24. Develop a logical argument
Q25. Write documents in discipline-appropriate style and format
Q26. Work effectively with others
Q27. Prepare and give oral presentations

```{r SALG-presently-Ican, results= TRUE}

# Q20. Find articles relevant to a particular problem in professional journals or elsewhere
# Q20 _ not sig
t.test(pre.post$Q20.27_1 ~ pre.post$type)
boxplot(pre.post$Q20.27_1 ~ pre.post$type)

# Q21. Critically read articles about issues raised in class
# Q21 _ not sig
t.test(pre.post$Q20.27_2 ~ pre.post$type)
boxplot(pre.post$Q20.27_2 ~ pre.post$type)

# Q22. Identify patterns in data
# Q22 _ not sig - same pre/post
t.test(pre.post$Q20.27_3 ~ pre.post$type)
boxplot(pre.post$Q20.27_3 ~ pre.post$type)

# Q23. Recognize a sound argument and appropriate use of evidence
# Q23 _ not sig
t.test(pre.post$Q20.27_4 ~ pre.post$type)
boxplot(pre.post$Q20.27_4 ~ pre.post$type)

# Q24. Develop a logical argument
# Q24 _ not sig
t.test(pre.post$Q20.27_5 ~ pre.post$type)
boxplot(pre.post$Q20.27_5 ~ pre.post$type)

# Q25. Write documents in discipline-appropriate style and format
# Q25 _ not sig
t.test(pre.post$Q20.27_6 ~ pre.post$type)
boxplot(pre.post$Q20.27_6 ~ pre.post$type)

# Q26. Work effectively with others
# Q26 _ not sig
t.test(pre.post$Q20.27_7 ~ pre.post$type)
boxplot(pre.post$Q20.27_7 ~ pre.post$type)

# Q27. Prepare and give oral presentations
# Q27 _ not sig
t.test(pre.post$Q20.27_8 ~ pre.post$type)
boxplot(pre.post$Q20.27_8 ~ pre.post$type)

```


(SALG) Presently I Understand…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q28. How ideas we will explore in this class relate to ideas I have encountered in other classes within this subject area
Q29. How ideas we will explore in this class relate to ideas I have encountered in classes outside of this subject area
Q30. How studying this subject helps people address real world issues


```{r SALG-i-understand, results = TRUE}
# Q28. How ideas we will explore in this class relate to ideas I have encountered in other classes within this subject area
# Q28 _ not sig
t.test(pre.post$Q28.30_1 ~ pre.post$type)
boxplot(pre.post$Q28.30_1 ~ pre.post$type)

# Q29. How ideas we will explore in this class relate to ideas I have encountered in classes outside of this subject area
# Q29 _ significant- but the reverse as expected
t.test(pre.post$Q28.30_2 ~ pre.post$type)
boxplot(pre.post$Q28.30_2 ~ pre.post$type)

# Q30. How studying this subject helps people address real world issues
# Q30 _ not sig - but could be without single outlier??
t.test(pre.post$Q28.30_3 ~ pre.post$type)
boxplot(pre.post$Q28.30_3 ~ pre.post$type)

```


(SALG) Presently I Am…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q31. Enthusiastic about this subject
Q32.  Interested in discussing this subject area with friends or family
Q33. Interested in taking or planning to take additional classes in this subject


```{r SALG-I-am, results = TRUE}

# Q31. Enthusiastic about this subject
# Q31 _ not sig - but a lot less post test variation
t.test(pre.post$Q31.33_1 ~ pre.post$type)
boxplot(pre.post$Q31.33_1 ~ pre.post$type)

# Q32.  Interested in discussing this subject area with friends or family
# Q32 _ not sig
t.test(pre.post$Q31.33_2 ~ pre.post$type)
boxplot(pre.post$Q31.33_2 ~ pre.post$type)


# Q33. Interested in taking or planning to take additional classes in this subject
# Q33 _ significant - but the reverse as expected
t.test(pre.post$Q31.33_3 ~ pre.post$type)
boxplot(pre.post$Q31.33_3 ~ pre.post$type)

```

Community Engagement
Rankings: Strongly agree, Agree, Neither agree nor disagree, disagree, strongly 
disagree, don’t know, prefer not to respond
Q34. I know what kinds of foods are grown agriculturally in my community
Q35. I know of areas where food can be foraged in or near the city I live
Q36. I know what kinds of tree species are found in my community
Q37. I know what kinds of diseases local foresters and agriculturalists manage in Boulder
Q38. I have met people who work in urban agriculture systems
Q39. I could tell a story about the environment at a place in or near the city I live
Q40. I could tell a story about the history of a place in or near the city I live

```{r comm-engagement, results = TRUE}

calcSE <- function(x){sd(x)/sqrt(length(x))}

# Q34. I know what kinds of foods are grown agriculturally in my community
#Q34 - 0.07- not signficant
t.test(pre.post$Q34.40_1 ~ pre.post$type)
boxplot(pre.post$Q34.40_1 ~ pre.post$type)
# Let's look at our group means and st errors
Q34.m <- aggregate(Q34.40_1~type, data=pre.post, mean)
Q34.m
Q34.sd <- aggregate(Q34.40_1~type, data=pre.post, calcSE)
Q34.sd

# Q35. I know of areas where food can be foraged in or near the city I live
#Q35 - significant
t.test(pre.post$Q34.40_2 ~ pre.post$type)
boxplot(pre.post$Q34.40_2 ~ pre.post$type)

# Q36. I know what kinds of tree species are found in my community
#Q36 - significant
t.test(pre.post$Q34.40_3 ~ pre.post$type)
boxplot(pre.post$Q34.40_3 ~ pre.post$type)

# Q37. I know what kinds of diseases local foresters and agriculturalists manage in Boulder
#Q37 - significant
t.test(pre.post$Q34.40_4 ~ pre.post$type)
boxplot(pre.post$Q34.40_4 ~ pre.post$type)

# Q38. I have met people who work in urban agriculture systems
#Q38 - significant
t.test(pre.post$Q34.40_5 ~ pre.post$type)
boxplot(pre.post$Q34.40_5 ~ pre.post$type)

# Q39. I could tell a story about the environment at a place in or near the city I live
#Q39 - significant
t.test(pre.post$Q34.40_6 ~ pre.post$type)
boxplot(pre.post$Q34.40_6 ~ pre.post$type)

# Q40. I could tell a story about the history of a place in or near the city I live
#Q40 - not significant
t.test(pre.post$Q34.40_7 ~ pre.post$type)
boxplot(pre.post$Q34.40_7 ~ pre.post$type)


```


################################################################################################

The SALG questions are not the same pre/post, so need to just look at histograms and compare before/after

Non-content questions
(SALG) Presently I Can…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q20. Find articles relevant to a particular problem in professional journals or elsewhere
Q21. Critically read articles about issues raised in class
Q22. Identify patterns in data
Q23. Recognize a sound argument and appropriate use of evidence
Q24. Develop a logical argument
Q25. Write documents in discipline-appropriate style and format
Q26. Work effectively with others
Q27. Prepare and give oral presentations

```{r SALG-presently-Ican, results= TRUE}

# Q20. Find articles relevant to a particular problem in professional journals or elsewhere
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_1, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q21. Critically read articles about issues raised in class
# some improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_2, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q22. Identify patterns in data
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_3, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q23. Recognize a sound argument and appropriate use of evidence
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_4, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q24. Develop a logical argument
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_5, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q25. Write documents in discipline-appropriate style and format
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_6, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q26. Work effectively with others
# minor improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_7, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q27. Prepare and give oral presentations
# improvement
ggplot(data = pre.post, mapping = aes(x=Q20.27_8, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

```


(SALG) Presently I Understand…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q28. How ideas we will explore in this class relate to ideas I have encountered in other classes within this subject area
Q29. How ideas we will explore in this class relate to ideas I have encountered in classes outside of this subject area
Q30. How studying this subject helps people address real world issues


```{r SALG-i-understand, results = TRUE}
# Q28. How ideas we will explore in this class relate to ideas I have encountered in other classes within this subject area
# no/minor improvment
ggplot(data = pre.post, mapping = aes(x=Q28.30_1, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q29. How ideas we will explore in this class relate to ideas I have encountered in classes outside of this subject area
# less consensus
ggplot(data = pre.post, mapping = aes(x=Q28.30_2, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

# Q30. How studying this subject helps people address real world issues
# improvement
ggplot(data = pre.post, mapping = aes(x=Q28.30_3, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

```


(SALG) Presently I Am…
Scale Response Category 1 = I gained nothing/not at all 2 = I gained a little 3 = I gained somewhat 4 = I gained a lot 5 = I gained a great deal [different scale than Pre-Assessment]
Q31. Enthusiastic about this subject
Q32.  Interested in discussing this subject area with friends or family
Q33. Interested in taking or planning to take additional classes in this subject


```{r SALG-I-am, results = TRUE}

# Q31. Enthusiastic about this subject
# overall gained
ggplot(data = pre.post, mapping = aes(x=Q31.33_1, color = type)) +
  geom_histogram() +
  facet_wrap(~type)


# Q32.  Interested in discussing this subject area with friends or family
# imrproved
ggplot(data = pre.post, mapping = aes(x=Q31.33_2, color = type)) +
  geom_histogram() +
  facet_wrap(~type)


# Q33. Interested in taking or planning to take additional classes in this subject
# overall gains
ggplot(data = pre.post, mapping = aes(x=Q31.33_3, color = type)) +
  geom_histogram() +
  facet_wrap(~type)

```
